Move settings profile editing to a dedicated edit page
Restyle the settings back button to the floating-bubble look of the group
info page and add an Edit button that opens a new /settings/edit route where
image, name, and reference location are edited (with a header Save + discard
confirmation), mirroring the /info → /edit-title pattern. The main settings
page now shows these read-only. Drop the four redundant helper captions and
promote Sign out to a separate full-width button behind a confirmation modal.
