+++ app/template.tsx
@@ -2,7 +2,6 @@
 
 import React, { useEffect, useState, Suspense } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
-import HeaderPortal from '@/components/HeaderPortal';
 import { useLongPress } from '@/lib/useLongPress';
 import { installClientLogForwarder } from '@/lib/clientLogForwarder';
 import { usePrefetch } from '@/lib/prefetch';
@@ -133,6 +132,9 @@ function TemplateInner({ children }: AppTemplateProps) {
   // of the new group button + padding treatment via isGroupRootView.
   const isGroupLikePage = isGroupRootView(pathname);
   const isSettingsPage = pathname === '/settings' || pathname === '/settings/';
+  // The profile editor (/settings/edit) renders its own fixed back + Save
+  // buttons via HeaderPortal, so it must opt out of the fallback header.
+  const isSettingsEditPage = pathname === '/settings/edit' || pathname === '/settings/edit/';
   // Phase G: /invite/<token> is a redemption landing page that renders
   // its own full-screen redirect-or-sign-in UI. The template's
   // fallback header would just sit above it as empty chrome.
@@ -148,7 +150,7 @@ function TemplateInner({ children }: AppTemplateProps) {
   return (
     <>
       {/* Fallback header for pages without a page-specific header (not group, settings, home, invite redemption, or create-modal). */}
-      {!isGroupFamilyPage && !isSettingsPage && !isInvitePage && pathname !== '/' && (
+      {!isGroupFamilyPage && !isSettingsPage && !isSettingsEditPage && !isInvitePage && pathname !== '/' && (
         <div className="sticky top-0 z-30 bg-white dark:bg-gray-900 border-b border-gray-200 dark:border-gray-700"
              style={{ paddingTop: 'env(safe-area-inset-top)' }}>
           <div className="relative flex items-start justify-between pt-2 pb-2 pl-2 pr-2.5">
@@ -241,24 +243,6 @@ function TemplateInner({ children }: AppTemplateProps) {
         </div>
       </div>
 
-      {/* Header elements rendered outside scaling container */}
-      <HeaderPortal>
-        {/* Back arrow in upper left — settings page always navigates to home. */}
-        {isSettingsPage && (
-          <div className="fixed left-0 z-50" style={{ top: 'calc(env(safe-area-inset-top, 0px) + 10px)' }}>
-            <button
-              onClick={() => navigateWithTransition(router, '/', 'back')}
-              className="w-12 h-16 flex items-center justify-center hover:bg-gray-100 dark:hover:bg-gray-800 rounded-full transition-colors"
-              aria-label="Go back"
-            >
-              <svg className="w-7 h-7 text-gray-600 dark:text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" />
-              </svg>
-            </button>
-          </div>
-        )}
-
-      </HeaderPortal>
     </>
   );
 }
